+++ permyt/mixins/requests/__init__.py
@@ -1,4 +1,5 @@
 from .connect import UserConnectMixin
+from .disconnect import UserDisconnectMixin
 from .provider import ProviderMixin
 from .requester import RequesterMixin
 from .scopes import ScopeManagementMixin
@@ -8,6 +9,7 @@ __all__ = (
     "RequesterMixin",
     "ProviderMixin",
     "UserConnectMixin",
+    "UserDisconnectMixin",
     "ScopeManagementMixin",
     "InboundMixin",
 )